pta: qcom: pas: refactor kodiak platform into platform-specific directory
The PAS authentication and boot sequences are difficult to abstract
across platforms: the original design relied on register map-based
abstractions to share logic, but the per-platform differences in
sequencing make that approach inadequate.

Move all Kodiak-specific logic (compute, lpass, q6dsp, video) and
the PTA command handlers into platform/kodiak/, and introduce a thin
dispatch interface (platform_pas.h) in pta_qcom_pas.c. Extract shared
resource table construction into platform/resource_table.c.
