+++ core/pta/qcom/pas/platform/kodiak/sub.mk
@@ -1,0 +1,3 @@
+srcs-y += entry.c compute.c dsp.c lpass.c rtable.c q6dsp.c video.c
+incdirs-y += .
+incdirs-y += ../
