+++ core/pta/qcom/pas/platform/kodiak/video.h
@@ -6,7 +6,13 @@
 #ifndef _VIDEO_H_
 #define _VIDEO_H_
 
+#include "pas_data.h"
+
 #define IRIS_WRAPPER_TOP_TZ_REG_BASE	0x000c0000
 #define IRIS_WRAPPER_TOP_REG_BASE	0x000b0000
 
-#endif
+TEE_Result venus_fw_start(struct qcom_pas_data *data);
+TEE_Result venus_fw_shutdown(struct qcom_pas_data *data);
+TEE_Result venus_fw_set_state(struct qcom_pas_data *data, bool power_on);
+
+#endif /* _VIDEO_H_ */
