+++ core/pta/qcom/pas/sub.mk
@@ -1,3 +1,4 @@
-srcs-y += pta_qcom_pas.c dsp.c pas.c q6dsp.c compute.c lpass.c venus.c
+srcs-y += pta_qcom_pas.c
 incdirs-y += .
-incdirs-y += platform/$(PLATFORM_FLAVOR)
+incdirs-y += platform/
+subdirs-y += platform/$(PLATFORM_FLAVOR)
